x11: uint -> ulong on clipboard read for x86-64

+++ gui-x11/x11.c
@@ -1426,7 +1426,7 @@ _xgetsnarf(XDisplay *xd)
 	uchar *data, *xdata;
 	Atom clipboard, type, prop;
 	unsigned long lastlen;
-	uint dummy, len;
+	unsigned long dummy, len;
 	int fmt, i;
 	Window w;
